preprocessor class draft
